review 141, 26, 344, 345, 88

+++ leetcode/344.reverse-string.cpp
@@ -1,8 +1,10 @@
 class Solution {
 public:
     string reverseString(string s) {
+        
         int start = 0;
-        int end = s.size() - 1;
+        int end = (int)(s.size() - 1);
+
         while(start < end)
         {
             char tmp = s[start];
@@ -12,7 +14,7 @@ public:
             start++;
             end--;
         }
-
+        
         return s;
     }
 };
